+++ R/checkArgumentsPDTMT.R
@@ -12,7 +12,7 @@
     stringIdCol, modificationsCol, excludeUnmodifiedPeptides,
     keepModifications, iColPattern, sampleAnnot, includeOnlySamples, excludeSamples,
     minScore, minDeltaScore, minPeptides, minPSMs, masterProteinsOnly,
-    imputeMethod, mergeGroups,
+    imputeMethod, assaysToExport, mergeGroups,
     comparisons, ctrlGroup, allPairwiseComparisons, singleFit,
     subtractBaseline, baselineGroup, normMethod, spikeFeatures, stattest,
     minNbrValidValues, minlFC, samSignificance, nperm, volcanoAdjPvalThr,
@@ -150,6 +150,7 @@
     ## Method choices
     .assertScalar(x = imputeMethod, type = "character",
                   validValues = c("impSeqRob", "MinProb"))
+    .assertVector(x = assaysToExport, type = "character", allowNULL = TRUE)
     .assertScalar(x = normMethod, type = "character",
                   validValues = c(MsCoreUtils::normalizeMethods(), "none"))
     .assertVector(x = spikeFeatures, type = "character", allowNULL = TRUE)