+++ R/runMaxQuantAnalysis.R
@@ -56,6 +56,13 @@
 #'     number of peptides is desired.
 #' @param imputeMethod Character string defining the imputation method to use.
 #'     Currently, \code{"impSeqRob"} and \code{"MinProb"} are supported.
+#' @param assaysToExport Character vector defining the name(s) of the assays
+#'     to use for exported abundances and barplots. This could, for example,
+#'     be set to an assay containing 'absolute' abundances, if available, even
+#'     if another assay is used for the actual analysis and comparison of
+#'     groups. If set to \code{NULL} or an assay name that does not exist in
+#'     the SingleCellExperiment object, the 'main' assay (defined by
+#'     \code{iColPattern}) will be used.
 #' @param mergeGroups Named list of character vectors defining sample groups
 #'     to merge to create new groups, that will be used for comparisons.
 #'     Any specification of \code{comparisons} or \code{ctrlGroup} should
@@ -229,7 +236,7 @@ runMaxQuantAnalysis <- function(
     iColPattern, sampleAnnot,
     includeOnlySamples = "", excludeSamples = "",
     minScore = 10, minPeptides = 2, imputeMethod = "MinProb",
-    mergeGroups = list(), comparisons = list(),
+    assaysToExport = c("iBAQ", "Top3"), mergeGroups = list(), comparisons = list(),
     ctrlGroup = "", allPairwiseComparisons = TRUE, singleFit = FALSE,
     subtractBaseline = FALSE, baselineGroup = "", normMethod = "none",
     spikeFeatures = NULL, stattest = "limma", minNbrValidValues = 2,
@@ -274,9 +281,9 @@ runMaxQuantAnalysis <- function(
         proteinIdCol = proteinIdCol, stringIdCol = stringIdCol,
         iColPattern = iColPattern, sampleAnnot = sampleAnnot,
         includeOnlySamples = includeOnlySamples,
-        excludeSamples = excludeSamples,
-        minScore = minScore, minPeptides = minPeptides,
-        imputeMethod = imputeMethod, mergeGroups = mergeGroups,
+        excludeSamples = excludeSamples, minScore = minScore,
+        minPeptides = minPeptides, imputeMethod = imputeMethod,
+        assaysToExport = assaysToExport, mergeGroups = mergeGroups,
         comparisons = comparisons, ctrlGroup = ctrlGroup,
         allPairwiseComparisons = allPairwiseComparisons, singleFit = singleFit,
         subtractBaseline = subtractBaseline, baselineGroup = baselineGroup,
@@ -318,9 +325,9 @@ runMaxQuantAnalysis <- function(
              reportTitle = reportTitle, reportAuthor = reportAuthor,
              iColPattern = iColPattern, sampleAnnot = sampleAnnot,
              includeOnlySamples = includeOnlySamples,
-             excludeSamples = excludeSamples,
-             minScore = minScore, minPeptides = minPeptides,
-             imputeMethod = imputeMethod, mergeGroups = mergeGroups,
+             excludeSamples = excludeSamples, minScore = minScore,
+             minPeptides = minPeptides, imputeMethod = imputeMethod,
+             assaysToExport = assaysToExport, mergeGroups = mergeGroups,
              comparisons = comparisons, ctrlGroup = ctrlGroup,
              allPairwiseComparisons = allPairwiseComparisons,
              singleFit = singleFit,