+++ tests/testthat/test-runMaxQuantAnalysis.R
@@ -39,6 +39,7 @@ test_that("runMaxQuantAnalysis works", {
         minScore = 10,
         minPeptides = 2,
         imputeMethod = "MinProb",
+        assaysToExport = c("iBAQ", "Top3"),
         mergeGroups = list(),
         comparisons = list(),
         ctrlGroup = "",
@@ -300,6 +301,12 @@ test_that("runMaxQuantAnalysis works", {
     args$imputeMethod <- "wrong"
     expect_error(do.call(runMaxQuantAnalysis, args),
                  "All values in 'imputeMethod' must be one of")
+
+    ## assaysToExport
+    args <- args0
+    args$assaysToExport <- 1
+    expect_error(do.call(runMaxQuantAnalysis, args),
+                 "'assaysToExport' must be of class 'character'")
 
     ## mergeGroups
     args <- args0